♻️ standard Pascal fucntions in a separate file
